fix: implement image deletion
