+++ features/stories/services/coverImage.js
@@ -1,0 +1,20 @@
+﻿import { backendUrl } from '../../../store/contantsDataNarrator';
+
+export function uploadCoverImage(storyId, coverImage) {
+  const formData = new FormData();
+  formData.append('files', coverImage);
+
+  return fetch(`${backendUrl}/stories/${storyId}/cover`, {
+    method: 'POST',
+    headers: {
+      'Content-Type': null,
+    },
+    body: formData
+  });
+}
+
+export function deleteCoverImage(storyId) {
+  return fetch(`${backendUrl}/stories/${storyId}/cover`, {
+    method: 'DELETE',
+  });
+}
